prefer dev requirements in setup.py
