+++ .travis.yml
@@ -6,5 +6,5 @@ python:
   - "2.7"
 
 script:
-  - pip install -r test_requirements.txt
+  - pip install --editable .[dev]
   - py.test ./test -s -vv
