+++ wampy/testing/pytest_plugin.py
@@ -8,7 +8,6 @@ import sys
 from contextlib import closing
 from time import time as now, sleep
 
-import colorlog
 import pytest
 
 from wampy.constants import DEFAULT_HOST, DEFAULT_PORT
@@ -19,17 +18,8 @@ from wampy.errors import ConnectionError
 
 logger = logging.getLogger('wampy.testing')
 
-logging_level_map = {
-    'DEBUG': logging.DEBUG,
-    'INFO': logging.INFO,
-}
-
 
 class ConfigurationError(Exception):
-    pass
-
-
-class PytestConfigurationError(Exception):
     pass
 
 
@@ -41,73 +31,6 @@ class TCPConnection(object):
     def connect(self):
         _socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
         _socket.connect((self.host, self.port))
-
-
-def add_file_logging():
-    root = logging.getLogger()
-    fhandler = logging.FileHandler(filename='test-runner-log.log', mode='a')
-    formatter = logging.Formatter(
-        '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
-    )
-    fhandler.setFormatter(formatter)
-    root.addHandler(fhandler)
-    root.setLevel(logging.DEBUG)
-
-
-def pytest_configure(config):
-    if config.option.logging_level is None:
-        logging_level = logging.INFO
-    else:
-        logging_level = config.option.logging_level
-        if logging_level not in logging_level_map:
-            raise PytestConfigurationError(
-                '{} not a recognised logging level'.format(logging_level)
-            )
-        logging_level = logging_level_map[logging_level]
-
-    sh = colorlog.StreamHandler()
-    sh.setLevel(logging_level)
-    formatter = colorlog.ColoredFormatter(
-        "%(white)s%(name)s %(reset)s %(log_color)s%"
-        "(levelname)-8s%(reset)s %(blue)s%(message)s",
-        datefmt=None,
-        reset=True,
-        log_colors={
-            'DEBUG': 'cyan',
-            'INFO': 'green',
-            'WARNING': 'yellow',
-            'ERROR': 'red',
-            'CRITICAL': 'red,bg_white',
-        },
-        secondary_log_colors={},
-        style='%'
-        )
-
-    sh.setFormatter(formatter)
-    root = logging.getLogger()
-    root.addHandler(sh)
-
-    if config.option.file_logging is True:
-        add_file_logging()
-
-
-def pytest_addoption(parser):
-    parser.addoption(
-        '--logging-level',
-        type=str,
-        action='store',
-        dest='logging_level',
-        help='configure the logging level',
-    )
-
-    parser.addoption(
-        '--file-logging',
-        type=bool,
-        action='store',
-        dest='file_logging',
-        help='optionally log to file',
-        default=False,
-    )
 
 
 class Crossbar(object):
